+++ packages/fluent-theme/package.json
@@ -30,9 +30,7 @@
     "@uifabric/tslint-rules": "^1.0.2"
   },
   "dependencies": {
-    "@uifabric/merge-styles": "^6.17.3",
     "@uifabric/set-version": "^1.1.3",
-    "@uifabric/styling": "^6.47.6",
     "@uifabric/variants": "^6.14.2",
     "office-ui-fabric-react": "^6.183.0",
     "tslib": "^1.7.1"
