+++ packages/fluent-theme/src/fluent/styles/Checkbox.styles.ts
@@ -1,4 +1,4 @@
-import { ICheckboxStyleProps, ICheckboxStyles } from 'office-ui-fabric-react/lib/Checkbox';
+import { ICheckboxStyleProps, ICheckboxStyles } from 'office-ui-fabric-react';
 
 export const CheckboxStyles = (props: ICheckboxStyleProps): Partial<ICheckboxStyles> => {
   const { disabled, checked, theme } = props;
