+++ packages/fluent-theme/src/fluent/styles/DetailsColumn.styles.ts
@@ -1,6 +1,5 @@
 import { FontSizes } from '../FluentType';
-import { FontWeights } from '@uifabric/styling';
-import { IDetailsColumnStyles } from 'office-ui-fabric-react/lib/components/DetailsList/DetailsColumn.types';
+import { FontWeights, IDetailsColumnStyles } from 'office-ui-fabric-react';
 
 export const DetailsColumnStyles = (): Partial<IDetailsColumnStyles> => {
   return {
