+++ packages/fluent-theme/src/fluent/styles/HoverCard.styles.ts
@@ -1,9 +1,4 @@
-import {
-  IExpandingCardStyleProps,
-  IExpandingCardStyles,
-  IPlainCardStyleProps,
-  IPlainCardStyles
-} from 'office-ui-fabric-react/lib/HoverCard';
+import { IExpandingCardStyleProps, IExpandingCardStyles, IPlainCardStyleProps, IPlainCardStyles } from 'office-ui-fabric-react';
 
 const commonCardStyles = (props: IExpandingCardStyleProps | IPlainCardStyleProps) => {
   const { theme } = props;
