+++ packages/fluent-theme/src/fluent/styles/Link.styles.ts
@@ -1,4 +1,4 @@
-import { ILinkStyleProps, ILinkStyles } from 'office-ui-fabric-react/lib/Link';
+import { ILinkStyleProps, ILinkStyles } from 'office-ui-fabric-react';
 
 export const LinkStyles = (props: ILinkStyleProps): Partial<ILinkStyles> => {
   const { isDisabled } = props;
