+++ packages/fluent-theme/src/fluent/styles/Persona.styles.ts
@@ -1,6 +1,5 @@
-import { IPersonaStyleProps, IPersonaStyles, PersonaSize, sizeBoolean } from 'office-ui-fabric-react/lib/Persona';
+import { IPersonaStyleProps, IPersonaStyles, PersonaSize, sizeBoolean, FontWeights } from 'office-ui-fabric-react';
 import { FontSizes } from '../FluentType';
-import { FontWeights } from 'office-ui-fabric-react/lib/Styling';
 
 export const PersonaStyles = (props: IPersonaStyleProps): Partial<IPersonaStyles> => {
   const size = sizeBoolean(props.size as PersonaSize);
