+++ packages/fluent-theme/src/fluent/styles/PrimaryButton.styles.ts
@@ -1,4 +1,4 @@
-import { IButtonStyles, IButtonProps } from 'office-ui-fabric-react/lib/Button';
+import { IButtonStyles, IButtonProps } from 'office-ui-fabric-react';
 
 export const PrimaryButtonStyles = (props: IButtonProps): Partial<IButtonStyles> => {
   const { theme } = props;
