+++ packages/fluent-theme/src/fluent/styles/Toggle.styles.ts
@@ -1,5 +1,4 @@
-import { FontWeights } from 'office-ui-fabric-react/lib/Styling';
-import { IToggleStyleProps, IToggleStyles } from 'office-ui-fabric-react/lib/Toggle';
+import { FontWeights, IToggleStyleProps, IToggleStyles } from 'office-ui-fabric-react';
 
 export const ToggleStyles = (props: IToggleStyleProps): Partial<IToggleStyles> => {
   const { disabled, checked, theme } = props;
