+++ packages/office-ui-fabric-react/src/Utilities.ts
@@ -1,2 +1,3 @@
 import './version';
 export * from '@uifabric/utilities';
+export { RectangleEdge } from './utilities/positioning/index';
